Fix varous minor bugs, compiler warnings and other tidy tasks. 200 Tests!

+++ E6502Lib/src/instructions/instruction_utils.h
@@ -39,7 +39,7 @@ namespace E6502 {
 				default: {
 					// TODO error handling
 					fprintf(stderr, "Invalid instruction provided to getRegFromInstruction: %x", instruction);
-					return 0x255;
+					return (Byte)0x255;
 				}
 			}
 		}
